moved WindowPtr from UiTasklet to TAsklet since it'S needed to access the gl::Context from Lua and I didn'T want to cast the Lua object in the script, changed uint8_t to uint32_t in createFrameBuffer since luabind still doesn't cope well with uint8_t, added more XXXPtrs, removed obsolete gl::State Lua binding

+++ projects/lostengine/lost/gl/FrameBuffer.h
@@ -35,7 +35,7 @@ namespace lost
       static TexturePtr createColorTexture(const math::Vec2& size, // size in pixels, result might vary
                                            GLenum internalFormat); // format of data as stored on gfx card
       static TexturePtr createDepthTexture(const math::Vec2& size, // size in pixels, result might vary
-                                           uint8_t bitDepth);      // desired bit depth, throws if bitdepth can'T be mapped to enum
+                                           uint32_t bitDepth);      // desired bit depth, throws if bitdepth can'T be mapped to enum
 
       /** creates a FBO with texture attachments for color 0 and depth.
        *
@@ -51,7 +51,7 @@ namespace lost
       static FrameBufferPtr createFrameBuffer(ContextPtr& ctx,
                                               const math::Vec2& size,
                                               GLenum colorFormat=GL_RGBA, // internal format of color attachment
-                                              uint8_t bitDepth=24); // desired bitdepth of depth attachment
+                                              uint32_t bitDepth=24); // desired bitdepth of depth attachment
                                                                  // lower resolutions will be tried if 
 
       FrameBuffer();
